삭제후 replace, history.back을 이용해 페이지 이동

+++ src/main/java/com/sbs/exam/demo/controller/UsrArticleController.java
@@ -102,7 +102,7 @@ public class UsrArticleController {
 
 	@RequestMapping("/usr/article/doDelete")
 	@ResponseBody
-	public ResultData<Integer> doDelete(HttpSession httpSession, int id) {
+	public String doDelete(HttpSession httpSession, int id) {
 		boolean isLogined = false;
 		int loginedMemberId = 0;
 
@@ -112,18 +112,20 @@ public class UsrArticleController {
 		}
 
 		if (isLogined == false) {
-			return ResultData.from("F-A", "로그인 후 이용해주세요");
+			return Ut.jsHistoryBack("로그인 후 이용해주세요");
 		}
 		
 		Article article = articleService.getForPrintArticle(loginedMemberId, id);
 
 		if (article == null) {
-			return ResultData.from("F-1", Ut.f("%d번 게시물은 존재하지 않습니다.", id));
+			return Ut.jsHistoryBack(Ut.f("%d번 게시물은 존재하지 않습니다.", id));
 		}
-		
+		if (article.getMemberId() != loginedMemberId) {
+			return Ut.jsHistoryBack("해당 게시물에 대한 권한이 없습니다");
+		}
 		articleService.deleteArticle(id);
 
-		return ResultData.from("S-1", Ut.f("%d번 게시물을 삭제했습니다.", id),"id", id);
+		return Ut.jsReplace(Ut.f("%d번 게시물을 삭제했습니다.", id), "/usr/article/list");
 	}
 
 	@RequestMapping("/usr/article/doModify")
@@ -142,14 +144,11 @@ public class UsrArticleController {
 		}
 		
 		Article article = articleService.getForPrintArticle(loginedMemberId, id);
-
-		if (article.getMemberId() != loginedMemberId) {
-			return ResultData.from("F-2", "해당 게시물에 대한 권한이 없습니다");
-		}
 		
 		if (article == null) {
 			return ResultData.from("F-1", Ut.f("%d번 게시물은 존재하지 않습니다.", id));
 		}
+		
 		ResultData actorCanModifyRd = articleService.actorCanModify(loginedMemberId, article);
 
 		if (actorCanModifyRd.isFail()) {
